Eight

+++ App/Controllers/AdminPanel.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers;
 
+use App\Exceptions\DbException;
 use App\Models\Article;
 
 class AdminPanel extends Admin
@@ -17,39 +18,48 @@ class AdminPanel extends Admin
         $this->view->display(__DIR__ . '/../../template/index.php');
     }
 
+    /**
+     * @throws DbException
+     */
     protected function editArticle()
     {
         $this->view->article = Article::findById($_GET['id']);
-
-        if (isset($_POST['submit']) && isset($_POST['title']) && isset($_POST['content'])) {
-            $this->view->article->title = $_POST['title'];
-            $this->view->article->content = $_POST['content'];
-            $this->view->article->save();
-            header('Location: index.php?ctrl=AdminPanel&action=showAllNews');
+        if (false !== Article::findById($_GET['id'])) {
+            if (isset($_POST['submit']) && isset($_POST['title']) && isset($_POST['content'])) {
+                $this->view->article->fill(['title' => $_POST['title'], 'content'=>$_POST['content']]);
+                $this->view->article->save();
+                header('Location: index.php?ctrl=AdminPanel&action=showAllNews');
+            } else {
+                $this->view->display(__DIR__ . '/../../template/editArticle.php');
+            }
         } else {
-            $this->view->display(__DIR__ . '/../../template/editArticle.php');
+            throw new DbException('Ошибка 404 - не найдено');
         }
     }
 
+    /**
+     * @throws DbException
+     */
     protected function deleteArticle()
     {
         $this->view->article = Article::findById($_GET['id']);
-
-        if (isset($_POST['submit'])) {
-            $this->view->article->delete();
-            header('Location: index.php?ctrl=AdminPanel&action=showAllNews');
+        if (false !== Article::findById($_GET['id'])) {
+            if (isset($_POST['submit'])) {
+                $this->view->article->delete();
+                header('Location: index.php?ctrl=AdminPanel&action=showAllNews');
+            } else {
+                $this->view->display(__DIR__ . '/../../template/deleteArticle.php');
+            }
         } else {
-            $this->view->display(__DIR__ . '/../../template/deleteArticle.php');
+            throw new DbException('Ошибка 404 - не найдено');
         }
     }
 
     protected function addArticle()
     {
         $this->view->article = new Article();
-
         if (isset($_POST['title']) && isset($_POST['content'])) {
-            $this->view->article->title = $_POST['title'];
-            $this->view->article->content = $_POST['content'];
+            $this->view->article->fill(['title' => $_POST['title'], 'content'=>$_POST['content']]);
             $this->view->article->save();
             header('Location: index.php?ctrl=AdminPanel&action=showAllNews');
         } else {
